fix print type

+++ QuPerturbation1DPrecision/QuPerturbation1DPrecision.cpp
@@ -46,11 +46,11 @@ void test0(double p)
 		inv3.Compute();
 
 		auto geti = [&](double x) -> size_t {  return (size_t)((x - x0) / (x1 - x0) * n); };
-		printf("%5d | %10.2E %10.2E %10.2E\n", n,
+		printf("%5d | %10.2E %10.2E %10.2E\n", (int)n,
 			abs2(inv1.GetPsi()[geti(-10)]) - solver.GetR(),
 			abs2(inv2.GetPsi()[geti(-10)]) - solver.GetR(),
 			abs2(inv3.GetPsi()[geti(-10)]) - solver.GetR());
-		printf("%5d | %10.2E %10.2E %10.2E\n", n,
+		printf("%5d | %10.2E %10.2E %10.2E\n", (int)n,
 			inv1.GetR() - solver.GetR(),
 			inv2.GetR() - solver.GetR(),
 			inv3.GetR() - solver.GetR());
